Move secp256k1 instruction helper to the sdk
