+++ sdk/src/lib.rs
@@ -34,7 +34,7 @@ pub mod program_utils;
 pub mod pubkey;
 pub mod recent_blockhashes_account;
 pub mod rpc_port;
-pub mod secp256k1;
+pub mod secp256k1_instruction;
 pub mod shred_version;
 pub mod signature;
 pub mod signers;
